+++ src/execution/plugins/llm-execution-tools/mocked/MockedEchoLlmExecutionTools.ts
@@ -2,11 +2,9 @@ import { spaceTrim } from 'spacetrim';
 import type { Prompt } from '../../../../types/Prompt';
 import { getCurrentIsoDate } from '../../../../utils/getCurrentIsoDate';
 import type { CommonExecutionToolsOptions } from '../../../CommonExecutionToolsOptions';
-import type { AvailableModel } from '../../../LlmExecutionTools';
-import type { LlmExecutionTools } from '../../../LlmExecutionTools';
-import type { PromptChatResult } from '../../../PromptResult';
-import type { PromptCompletionResult } from '../../../PromptResult';
-import { addPromptResultUsage } from '../../../addPromptResultUsage';
+import type { AvailableModel, LlmExecutionTools } from '../../../LlmExecutionTools';
+import type { PromptChatResult, PromptCompletionResult } from '../../../PromptResult';
+import { addUsage } from '../../../addUsage';
 
 /**
  * Mocked execution Tools for just echoing the requests for testing purposes.
@@ -34,7 +32,7 @@ export class MockedEchoLlmExecutionTools implements LlmExecutionTools {
                 start: getCurrentIsoDate(),
                 complete: getCurrentIsoDate(),
             },
-            usage: addPromptResultUsage(/* <- TODO: [🧠] Compute here at least words, characters,... etc */),
+            usage: addUsage(/* <- TODO: [🧠] Compute here at least words, characters,... etc */),
             rawResponse: {
                 note: 'This is mocked echo',
             },
@@ -61,7 +59,7 @@ export class MockedEchoLlmExecutionTools implements LlmExecutionTools {
                 start: getCurrentIsoDate(),
                 complete: getCurrentIsoDate(),
             },
-            usage: addPromptResultUsage(/* <- TODO: [🧠] Compute here at least words, characters,... etc */),
+            usage: addUsage(/* <- TODO: [🧠] Compute here at least words, characters,... etc */),
             rawResponse: {
                 note: 'This is mocked echo',
             },